+++ addOns/oast/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.12.0] - 2022-10-19
 ### Fixed
 - Deregister the Interactsh service even in case of error (Issue 7504).
 - Clear Interactsh payloads from the GUI when the service is deregistered.
@@ -124,6 +124,7 @@ Button is disabled in that case.
 
 ## [0.1.0] - 2021-08-04
 
+[0.12.0]: https://github.com/zaproxy/zap-extensions/releases/oast-v0.12.0
 [0.11.0]: https://github.com/zaproxy/zap-extensions/releases/oast-v0.11.0
 [0.10.0]: https://github.com/zaproxy/zap-extensions/releases/oast-v0.10.0
 [0.9.0]: https://github.com/zaproxy/zap-extensions/releases/oast-v0.9.0